requirements.txt
flask==0.11
sqlalchemy==1.0.14
mysqlclient==1.3.7
paho-mqtt>=1.2
influxdb>=0.9
docker-py
requests